Вынесла общие методы для модалок в родительский класс, добавила page object для редактирования/создания заказа, которые наследуются от родительского
